recognize every logical plan and match the related resources and actions
