+++ core/src/main/scala/org/apache/spark/sql/crossdata/config/CoreConfig.scala
@@ -27,6 +27,8 @@ import scala.util.Try
 
 object CoreConfig {
 
+  val DefaultCatalogIdentifier = "defaultcat"
+
   val CoreBasicConfig = "core-reference.conf"
   val ParentConfigName = "crossdata-core"
   val CoreUserConfigFile = "external.config.filename"
@@ -44,18 +46,16 @@ object CoreConfig {
   val SecurityConfigKey = "security"
   val SecurityManagerConfigKey = "manager"
   val ClassConfigKey = "class"
+  val PrefixKey = "prefix"
 
   val AuditConfigKey = "audit"
   val UserConfigKey = "user"
   val PasswordConfigKey = "password"
   val SessionConfigKey = "session"
   val CatalogClassConfigKey = s"$CatalogConfigKey.$ClassConfigKey"
+  val CatalogPrefixConfigKey = s"$CatalogConfigKey.$PrefixKey" // TODO rename prefix to catalogIdentifier
   val StreamingCatalogClassConfigKey = s"$StreamingConfigKey.$CatalogConfigKey.$ClassConfigKey"
   val SecurityClassConfigKey = s"$SecurityConfigKey.$SecurityManagerConfigKey.$ClassConfigKey"
-  val SecurityAuditConfigKey = s"$SecurityConfigKey.$SecurityManagerConfigKey.$AuditConfigKey"
-  val SecurityUserConfigKey = s"$SecurityConfigKey.$SecurityManagerConfigKey.$UserConfigKey"
-  val SecurityPasswordConfigKey = s"$SecurityConfigKey.$SecurityManagerConfigKey.$PasswordConfigKey"
-  val SecuritySessionConfigKey = s"$SecurityConfigKey.$SecurityManagerConfigKey.$SessionConfigKey"
 
   val SparkSqlConfigPrefix = "config.spark.sql" //WARNING!! XDServer is using this path to read its parameters
